fix typos

+++ scripts/ahmed_body/pcno_ahmedbody_test.py
@@ -115,7 +115,7 @@ features /= np.array([1.0, 100.0, 100.0, 100.0, 100.0, 180.0/np.pi, 100.0, 100.0
 # keep only the pressure and Reynolds number
 features = torch.from_numpy(features.astype(np.float32))
 
-directed_edges = torch.from_numpy(directed_edges)
+directed_edges = torch.from_numpy(directed_edges.astype(np.int64))
 edge_gradient_weights = torch.from_numpy(
     edge_gradient_weights.astype(np.float32))
