+++ scripts/airfoil_flap/pcno_airfoilflap_test.py
@@ -120,7 +120,7 @@ nodes = torch.from_numpy(nodes.astype(np.float32))
 node_weights = torch.from_numpy(node_weights.astype(np.float32))
 node_rhos = torch.from_numpy(node_rhos.astype(np.float32))
 features = torch.from_numpy(features.astype(np.float32))
-directed_edges = torch.from_numpy(directed_edges)
+directed_edges = torch.from_numpy(directed_edges.astype(np.int64))
 edge_gradient_weights = torch.from_numpy(edge_gradient_weights.astype(np.float32))
 
 # This is important
